requirements.txt
black==21.6b0
discord==1.7.3
python-dotenv==0.18.0
